requirements.txt
chromadb==0.6.3
flet==0.26.0
flet-desktop-light==0.26.0
google-genai==1.12.1
google-generativeai==0.8.3
gradio==5.23.0
html2text==2024.2.26
jupyter==1.0.0
kaleido==0.2.1
langchain-core==0.3.33
langchain-community==0.3.16
langchain==0.3.17
langchain-experimental==0.3.4
langchain-google-genai==2.0.9
langchain-openai==0.3.2
langchain-chroma==0.2.2
llama-index==0.12.21
llama-index-embeddings-gemini==0.3.2
llama-index-embeddings-google-genai
llama-index-embeddings-huggingface==0.5.2
llama-index-finetuning==0.3.0
llama-index-llms-gemini==0.4.5
llama-index-llms-google-genai==0.1.7
llama-index-llms-together==0.3.1
llama-index-readers-web==0.3.7
llama-index-vector-stores-chroma==0.4.1
llama-index-vector-stores-qdrant==0.6.0
openai==1.59.8
pandas==2.2.3
pillow==10.4.0
pydantic==2.10.5
python-language-server==0.36.2
python-lsp-server==1.11.0
scikit-learn==1.6.1
sentence-transformers==3.4.1
linkedin-api-client==0.3.0
newspaper3k==0.2.8
lxml_html_clean==0.4.1
arxiv==2.1.3
pdf2image==1.17.0
poppler-utils==0.1.0
python-docx==1.1.2
fastembed==0.7.0
